requirements.txt
playwright==1.40.0
behave==1.2.6
pytest==7.4.3
pytest-playwright==0.4.3
behave-html-formatter==0.9.10
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
